+++ src/sll/array.c
@@ -256,9 +256,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_array_equal_map(const sll_array
 	}
 	for (sll_map_length_t i=0;i<(m->l<<1);i+=2){
 		sll_object_t* e=m->v[i];
-		GC_LOCK(e);
 		sll_bool_t st=(e->t!=SLL_OBJECT_TYPE_INT||e->dt.i<0||e->dt.i>=a->l||!sll_operator_strict_equal(m->v[i+1],a->v[e->dt.i]));
-		GC_UNLOCK(e);
 		if (st){
 			return 0;
 		}
@@ -547,7 +545,7 @@ __SLL_EXTERNAL void sll_array_remove(const sll_array_t* a,sll_object_t* v,sll_ar
 				i++;
 			}
 			else{
-				sll_release_object(o->v[j]);
+				GC_RELEASE(o->v[j]);
 			}
 		}
 		o->l=i;
@@ -754,7 +752,7 @@ __SLL_EXTERNAL void sll_array_split(const sll_array_t* a,sll_object_t* e,sll_arr
 __SLL_EXTERNAL void sll_array_set(const sll_array_t* a,sll_array_length_t i,sll_object_t* v){
 	if (i<a->l){
 		SLL_ACQUIRE(v);
-		sll_release_object(a->v[i]);
+		GC_RELEASE(a->v[i]);
 		a->v[i]=v;
 	}
 }
@@ -867,7 +865,7 @@ __SLL_EXTERNAL void sll_array_xor(const sll_array_t* a,const sll_array_t* b,sll_
 
 __SLL_EXTERNAL void sll_free_array(sll_array_t* a){
 	for (sll_array_length_t i=0;i<a->l;i++){
-		sll_release_object(a->v[i]);
+		GC_RELEASE(a->v[i]);
 	}
 	sll_allocator_release(a->v);
 	a->v=NULL;