+++ src/sll/cast.c
@@ -1,4 +1,5 @@
 #include <sll/_internal/common.h>
+#include <sll/_internal/gc.h>
 #include <sll/api/math.h>
 #include <sll/api/string.h>
 #include <sll/array.h>
@@ -84,7 +85,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_operator_cast(sll_object_t* 
 					if (dt->fn.str){
 						sll_object_t* v=sll_execute_function(dt->fn.str,&a,1,0);
 						sll_object_t* str=sll_operator_cast(v,sll_static_int[SLL_OBJECT_TYPE_STRING]);
-						sll_release_object(v);
+						GC_RELEASE(v);
 						return str;
 					}
 					break;
@@ -111,7 +112,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_operator_cast(sll_object_t* 
 		if (sll_current_runtime_data&&b->dt.i<=sll_current_runtime_data->tt->l+SLL_MAX_OBJECT_TYPE){
 			sll_object_t* src=sll_operator_cast(a,sll_static_int[SLL_OBJECT_TYPE_ARRAY]);
 			sll_object_t* o=sll_create_object_type(sll_current_runtime_data->tt,(sll_object_type_t)(b->dt.i),src->dt.a.v,src->dt.a.l);
-			sll_release_object(src);
+			GC_RELEASE(src);
 			return o;
 		}
 		SLL_ACQUIRE(a);