+++ src/sll/var_arg.c
@@ -1,3 +1,4 @@
+#include <sll/_internal/gc.h>
 #include <sll/common.h>
 #include <sll/gc.h>
 #include <sll/operator.h>
@@ -34,7 +35,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_char_t sll_var_arg_get_char(sll_var_arg_li
 	}
 	sll_object_t* n=sll_operator_cast((sll_object_t*)(*(va->dt.sll.p)),sll_static_int[SLL_OBJECT_TYPE_CHAR]);
 	sll_char_t o=n->dt.c;
-	sll_release_object(n);
+	GC_RELEASE(n);
 	va->dt.sll.p++;
 	va->dt.sll.l--;
 	return o;
@@ -51,7 +52,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_float_t sll_var_arg_get_float(sll_var_arg_
 	}
 	sll_object_t* n=sll_operator_cast((sll_object_t*)(*(va->dt.sll.p)),sll_static_int[SLL_OBJECT_TYPE_FLOAT]);
 	sll_float_t o=n->dt.f;
-	sll_release_object(n);
+	GC_RELEASE(n);
 	va->dt.sll.p++;
 	va->dt.sll.l--;
 	return o;
@@ -68,7 +69,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_integer_t sll_var_arg_get_int(sll_var_arg_
 	}
 	sll_object_t* n=sll_operator_cast((sll_object_t*)(*(va->dt.sll.p)),sll_static_int[SLL_OBJECT_TYPE_INT]);
 	sll_integer_t o=n->dt.i;
-	sll_release_object(n);
+	GC_RELEASE(n);
 	va->dt.sll.p++;
 	va->dt.sll.l--;
 	return o;
@@ -107,7 +108,7 @@ __SLL_EXTERNAL void sll_var_arg_get_string(sll_var_arg_list_t* va,sll_string_t* 
 	else{
 		sll_object_t* n=sll_operator_cast((sll_object_t*)(*(va->dt.sll.p)),sll_static_int[SLL_OBJECT_TYPE_STRING]);
 		sll_string_clone(&(n->dt.s),o);
-		sll_release_object(n);
+		GC_RELEASE(n);
 		va->dt.sll.p++;
 		va->dt.sll.l--;
 	}