+++ CHANGELOG.md
@@ -76,6 +76,7 @@
 - Deprecated macros: `SLL_OBJECT_GET_TYPE`, `SLL_OBJECT_GET_TYPE_MASK` and `SLL_OBJECT_IS_STORAGE`
 - Unneeded flag: `SLL_OBJECT_FLAG_CONSTANT`
 - Unused functions: `sll_object_get_type` and `sll_set_sandbox_flag_string`
+- Unneeded functions: `sll_lock_object` and `sll_unlock_object`
 
 ## [0.7.10] - 2022-03-26